requirements.txt
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0